Login with user object in key value pair of Async Storage
Here the problem was we cannot store object directly n key value pair, so we have to convert object in json string by using JSON.Stringify(objVariable), the object will be converted into a simple string which still looks like json but that the only string. \n Moreover when we get data from Async Storage using getItem, to use string as object, we have to parse the string, which is already in syntax of json object, into json object by using JSON.parse(string). It will convert the string (which we initially converted from json to string) into JSON object. Now we can use it in javascript as object