global - backword props - re-render component
global is predefined variable which can be used in anywhere in the react native app, there is no need to import global, u can simply use it \nbackword props is a challenge, so we used global variable. \n Here we re-rendered the component by changing values from next screen to effect on previous screen. So we add listner event and make a return call in useEffect to re-render the component if there is change in dependancy, i.e. navigation dependancy